+++ src/main/java/domain/Article.java
@@ -2,14 +2,9 @@ package domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 
-public class Article extends Reference {
-
-    public String author;
-    public String title;
-    public String journal;
-    public String year;
-    public String volume;
+public class Article implements Reference {
 
     private ArrayList<String> requiredFields;
     private HashMap<String, String> fields;
@@ -21,48 +16,40 @@ public class Article extends Reference {
     }
 //CHECKSTYLE:OFF
 
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public void setJournal(String journal) {
-        this.journal = journal;
-    }
-
-    public void setYear(String year) {
-        this.year = year;
-    }
-
-    public void setVolume(String volume) {
-        this.volume = volume;
-    }
-
+    @Override
     public boolean hasRequiredFields() {
-        if (this.author == null) {
-            return false;
-        } else if (this.title == null) {
-            return false;
-        } else if (this.journal == null) {
-            return false;
-        } else if (this.year == null) {
-            return false;
-        } else if (this.volume == null) {
-            return false;
+        for (String field : this.requiredFields) {
+            if (this.fields.get(field) == null) {
+                return false;
+            }
         }
         return true;
     }
 //CHECKSTYLE:ON
 
     @Override
-    void initRequiredFields() {
+    public void initRequiredFields() {
         requiredFields.add("author");
         requiredFields.add("title");
         requiredFields.add("journal");
         requiredFields.add("year");
         requiredFields.add("volume");
+    }
+
+    @Override
+    public void setField(String field, String value) {
+        this.fields.put(field, value);
+    }
+
+    @Override
+    public String getField(String field) {
+        if (this.fields.get(field) == null) {
+            return null;
+        } else return this.fields.get(field);
+    }
+    
+    @Override
+    public List getRequiredFields() {
+        return this.requiredFields;
     }
 }
