+++ src/main/java/domain/Inproceedings.java
@@ -2,14 +2,9 @@ package domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 
-public class Inproceedings extends Reference {
-    
-    public String author;
-    public String title;
-    public String booktitle;
-    public String year;
-
+public class Inproceedings implements Reference {
     private ArrayList<String> requiredFields;
     private HashMap<String, String> fields;
 
@@ -18,46 +13,41 @@ public class Inproceedings extends Reference {
         this.fields = new HashMap<>();
         initRequiredFields();
     }
-
 //CHECKSTYLE:OFF
-    public void setAuthor(String author) {
-        this.author = author;
-    }
 
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public void setBookTitle(String BookTitle) {
-        this.booktitle = BookTitle;
-    }
-
-    public void setYear(String year) {
-        this.year = year;
-    }
-
-    // @Override
+    @Override
     public boolean hasRequiredFields() {
-        if (this.author == null) {
-            return false;
-        } else if (this.title == null) {
-            return false;
-        } else if (this.booktitle == null) {
-            return false;
-        } else if (this.year == null) {
-            return false;
-        } else {
-            return true;
+        for (String field : this.requiredFields) {
+            if (this.fields.get(field) == null) {
+                return false;
+            }
         }
+        return true;
     }
 //CHECKSTYLE:ON
 
-    @Override
-    void initRequiredFields() {
+   @Override
+    public void initRequiredFields() {
         this.requiredFields.add("author");
         this.requiredFields.add("title");
         this.requiredFields.add("booktitle");
         this.requiredFields.add("year");
     }
 
+    @Override
+    public void setField(String field, String value) {
+        this.fields.put(field, value);
+    }
+
+    @Override
+    public String getField(String field) {
+        if (this.fields.get(field) == null) {
+            return null;
+        } else return this.fields.get(field);
+    }
+    
+    @Override
+    public List getRequiredFields() {
+        return this.requiredFields;
+    }
 }
